functions/requirements.txt
# Base requirements for core Firebase functions
# This is what gets installed in CI/CD

# === Firebase & Google Cloud ===
firebase-admin==6.5.0
firebase-functions==0.4.2
functions-framework==3.5.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-pubsub==2.20.1
google-cloud-secret-manager==2.20.2
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.1
grpcio==1.62.1
grpcio-status==1.62.1
proto-plus==1.23.0
protobuf==4.25.3

# === Web Framework ===
Flask==3.0.2
Flask-Cors==4.0.0
fastapi==0.100.0
gunicorn==21.2.0
Werkzeug==3.0.1
blinker==1.7.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5

# === Data Processing ===
pandas==2.2.1
numpy==1.26.4
pytz==2024.1
python-dateutil==2.9.0.post0
tzdata==2024.1
scikit-learn==1.7.0
scipy==1.14.1

# === Visualization (needed by sales_report) ===
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.17.0
mplsoccer==1.4.0
pillow==10.2.0
reportlab==4.1.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
kiwisolver==1.4.5

# === Image Processing (needed by processProductImage) ===
opencv-python==4.8.0.76

# === AWS Integration ===
boto3==1.36.18

# === File Processing ===
openpyxl==3.1.3
PyPDF2==3.0.1
et-xmlfile==1.1.0
fastparquet==2024.2.0
cramjam==2.8.2
xlrd==2.0.1

# === HTTP & Requests ===
requests==2.31.0
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
httplib2==0.22.0

# === Utilities ===
python-dotenv==1.0.1
python-Levenshtein==0.26.1
fuzzywuzzy==0.18.0
beautifulsoup4==4.12.3
chardet==5.2.0
schedule==1.2.1
holidays==0.36
watchdog==4.0.0

# === Authentication & Security ===
PyJWT==2.8.0
cryptography==42.0.5
cffi==1.16.0
pycparser==2.21
rsa==4.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
oauth2client==4.1.3

# === Caching & Storage ===
CacheControl==0.14.0
cachetools==5.3.3
msgpack==1.0.8

# === Other ===
cloudevents==1.9.0
deprecation==2.1.0
fsspec==2024.2.0
gcloud==0.18.3
packaging==24.0
pathlib==1.0.1
pyparsing==3.1.2
six==1.16.0
typing_extensions==4.10.0
uritemplate==4.1.1
PyYAML==6.0.1
